requirements.txt
streamlit
tensorflow==2.19.0
pillow
numpy
opencv-python-headless
matplotlib
pandas